+++ linbox/ring/ntl/ntl-lzz_px.h
@@ -582,7 +582,7 @@ namespace LinBox
 		typedef NTL::zz_pX Element;
 		UnparametricRandIter<NTL::zz_pX>(const NTL_zz_pX & F ,
 						 const size_t& size = 0,
-						 const size_t& seed = 0
+						 const uint64_t& seed = 0
 						) :
                 _size(size), _seed(seed), _ring(F)
 		{
@@ -621,7 +621,7 @@ namespace LinBox
 
 	protected:
 		size_t _size;
-		size_t _seed;
+		uint64_t _seed;
         const NTL_zz_pX& _ring;
 	}; // class UnparametricRandIters
